Vendor Ruby if an installation is not found
The ERB templating feature of this buildpack requires that the
`erb` command (part of Ruby) be available at runtime, in order
that the nginx config templates can be rendered into a valid nginx
config.

On Heroku-22 the stack image no longer includes a system Ruby
installation, so in order for this buildpack to continue to work, a
Ruby install must be vendored by this buildpack.

If an app already uses the Ruby buildpack, and that buildpack is
ordered prior to the nginx buildpack, then this buildpack will skip
the Ruby vendoring step to save installing a redundant copy of Ruby.

Fixes #101.
GUS-W-11321729.