Link against system PCRE
Absolutely no need to compile this ourselves.

Because we have always built a custom libpcre3 (v8.x) and not the more modern PCRE2 (v10.x), we are enforcing usage of the old version for now.

This is to ensure that existing configs with regexes continue to work, as PCRE2 is more aggressive in its pattern validation.

To give a simple example, `/[\w-.]+/` now throws "Compilation failed: invalid range in character class at offset 3", and the `-` needs to be escaped, or moved to the end of the character class.